+++ core/src/test/java/cucumber/runtime/HookOrderTest.java
@@ -23,14 +23,14 @@ import static org.mockito.Mockito.when;
 
 public class HookOrderTest {
 
-    private Runtime runtime;
+    private LegacyRuntime runtime;
     private Glue glue;
 
     @Before
     public void buildMockWorld() {
         ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
         RuntimeOptions runtimeOptions = new RuntimeOptions("");
-        runtime = new Runtime(mock(ResourceLoader.class), classLoader, asList(mock(Backend.class)), runtimeOptions);
+        runtime = new LegacyRuntime(mock(ResourceLoader.class), classLoader, asList(mock(Backend.class)), runtimeOptions);
         runtime.buildBackendWorlds(null, Collections.<Tag>emptySet(), mock(gherkin.formatter.model.Scenario.class));
         glue = runtime.getGlue();
     }
