+++ core/src/test/java/cucumber/runtime/TestHelper.java
@@ -4,7 +4,6 @@ import cucumber.api.PendingException;
 import cucumber.runtime.formatter.StepMatcher;
 import cucumber.runtime.io.ClasspathResourceLoader;
 import cucumber.runtime.io.Resource;
-import cucumber.runtime.model.CucumberFeature;
 import gherkin.I18n;
 import gherkin.formatter.Formatter;
 import gherkin.formatter.Reporter;
@@ -82,7 +81,7 @@ public class TestHelper {
         final ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
         final ClasspathResourceLoader resourceLoader = new ClasspathResourceLoader(classLoader);
         final RuntimeGlue glue = createMockedRuntimeGlueThatMatchesTheSteps(stepsToResult, stepsToLocation, hooks);
-        final Runtime runtime = new Runtime(resourceLoader, classLoader, asList(mock(Backend.class)), runtimeOptions, new StopWatch.Stub(stepHookDuration), glue);
+        final LegacyRuntime runtime = new LegacyRuntime(resourceLoader, classLoader, asList(mock(Backend.class)), runtimeOptions, new StopWatch.Stub(stepHookDuration), glue);
 
         for (CucumberFeature feature : features) {
             feature.run(formatter, reporter, runtime);
