+++ core/src/test/java/cucumber/runtime/formatter/JSONPrettyFormatterTest.java
@@ -2,6 +2,7 @@ package cucumber.runtime.formatter;
 
 import cucumber.runtime.Backend;
 import cucumber.runtime.HookDefinition;
+import cucumber.runtime.LegacyRuntime;
 import cucumber.runtime.Runtime;
 import cucumber.runtime.RuntimeOptions;
 import cucumber.runtime.StopWatch;
@@ -74,7 +75,7 @@ public class JSONPrettyFormatterTest {
         RuntimeOptions runtimeOptions = new RuntimeOptions(args);
         Backend backend = mock(Backend.class);
         when(backend.getSnippet(any(Step.class), any(FunctionNameGenerator.class))).thenReturn("TEST SNIPPET");
-        final Runtime runtime = new Runtime(resourceLoader, classLoader, asList(backend), runtimeOptions, new StopWatch.Stub(1234), null);
+        final Runtime runtime = new LegacyRuntime(resourceLoader, classLoader, asList(backend), runtimeOptions, new StopWatch.Stub(1234), null);
         runtime.getGlue().addBeforeHook(hook);
         runtime.run();
         return report;
