+++ core/src/test/java/cucumber/runtime/formatter/RerunFormatterTest.java
@@ -1,7 +1,7 @@
 package cucumber.runtime.formatter;
 
 import cucumber.runtime.TestHelper;
-import cucumber.runtime.model.CucumberFeature;
+import cucumber.runtime.CucumberFeature;
 import org.junit.Test;
 
 import java.util.AbstractMap.SimpleEntry;
