+++ junit/src/test/java/cucumber/runtime/junit/ExecutionUnitRunnerTest.java
@@ -1,19 +1,12 @@
 package cucumber.runtime.junit;
 
-import cucumber.runtime.FeatureBuilder;
 import cucumber.runtime.io.ClasspathResourceLoader;
-import cucumber.runtime.io.Resource;
-import cucumber.runtime.model.CucumberFeature;
-import cucumber.runtime.model.CucumberScenario;
+import cucumber.runtime.CucumberFeature;
+import cucumber.runtime.CucumberScenario;
 import gherkin.formatter.model.Step;
 import org.junit.Test;
 import org.junit.runner.Description;
 
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.UnsupportedEncodingException;
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
